Changes that are clear from @TeslaDethray's requests.
