+++ src/Terminus.php
@@ -143,12 +143,13 @@ class Terminus implements ConfigAwareInterface, ContainerAwareInterface, LoggerA
             'Pantheon\\Terminus\\Commands\\Connection\\InfoCommand',
             'Pantheon\\Terminus\\Commands\\Connection\\SetCommand',
             'Pantheon\\Terminus\\Commands\\Dashboard\\ViewCommand',
+            'Pantheon\\Terminus\\Commands\\Domain\\Primary\\SetCommand',
+            'Pantheon\\Terminus\\Commands\\Domain\\Primary\\RemoveCommand',
             'Pantheon\\Terminus\\Commands\\Domain\\AddCommand',
             'Pantheon\\Terminus\\Commands\\Domain\\DNSCommand',
             'Pantheon\\Terminus\\Commands\\Domain\\ListCommand',
             'Pantheon\\Terminus\\Commands\\Domain\\LookupCommand',
             'Pantheon\\Terminus\\Commands\\Domain\\RemoveCommand',
-            'Pantheon\\Terminus\\Commands\\Domain\\SetPrimaryCommand',
             'Pantheon\\Terminus\\Commands\\Env\\ClearCacheCommand',
             'Pantheon\\Terminus\\Commands\\Env\\CloneContentCommand',
             'Pantheon\\Terminus\\Commands\\Env\\CodeLogCommand',
